requirements.txt
pandas
numpy
tqdm
python-dotenv
pydantic
moviepy
requests
ipykernel
notebook
openai
langchain
langchain-community
chromadb
unstructured
groq
replicate
pypdf
huggingface_hub

